переделал обработчики ошибок при загрузке данных
